Windows System Performance Optimizer/requirements.txt
bcrypt==4.2.1
certifi==2024.12.14
charset-normalizer==3.4.0
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.10
packaging==24.2
pillow==11.0.0
psutil==6.1.0
pycryptodome==3.21.0
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
pywin32==308
requests==2.32.3
tk==0.1.0
ttkbootstrap==1.10.1
urllib3==2.2.3
winshell==0.6
WMI==1.5.1
